refactor: remove `/api/endpoints` because response payload is >5mb
