+++ lib/get-endpoints.js
@@ -3,16 +3,27 @@ module.exports = getEndpoints;
 const { filter } = require("lodash");
 const got = require("got");
 
-const getVercelUrl = require("./get-vercel-url");
+const getReleases = require("../lib/get-releases");
+const schemaToEndpoints = require("../lib/schema-to-endpoints");
 
 async function getEndpoints(options) {
-  const { body } = await got(`${getVercelUrl()}/api/endpoints`, {
-    searchParams: {
-      version: options.version,
-      ghe: options.ghe,
-    },
-  });
-  let result = JSON.parse(body);
+  const releases = await getReleases();
+  const version = options.version || releases[0].version;
+
+  const knownVersions = releases.map((release) => release.version).join(",");
+  if (!knownVersions.includes(version)) {
+    res.status(404);
+    res.json({
+      error: `Version "${version}" could not be found. Knonw versions are: ${knownVersions}`,
+    });
+    return;
+  }
+
+  const url = toJsonFileUrl(version, options.ghe);
+  console.log(`downloading ${url}`);
+  const { body } = await got(url);
+
+  let result = schemaToEndpoints(JSON.parse(body));
 
   if (options.filter) {
     result = filter(result, options.filter);
@@ -77,4 +88,11 @@ function sortByScopeAndId(a, b) {
   if (a.id < b.id) return -1;
 
   return 0;
+}
+
+function toJsonFileUrl(version, ghe) {
+  const baseUrl = `https://unpkg.com/@octokit/openapi@${version}/generated`;
+  if (!ghe) return `${baseUrl}/api.github.com.deref.json`;
+
+  return baseUrl + ghe.replace(/^GHE_(\d)(\d+)/, `/ghes-$1.$2$.deref.json`);
 }
